requirements.txt
Flask==2.0.3
Werkzeug==2.0.3
pymongo==3.12.0
python-dotenv==0.19.0